move utils for project
