+++ web/e2e/project/asset.spec.ts
@@ -1,6 +1,7 @@
 import { expect, test } from "@reearth-cms/e2e/utils";
 
-import { createProject, deleteProject, crudComment } from "./utils";
+import { crudComment } from "./utils";
+import { createProject, deleteProject } from "./utils/project";
 
 const uploadFileUrl =
   "https://assets.cms.plateau.reearth.io/assets/11/6d05db-ed47-4f88-b565-9eb385b1ebb0/13100_tokyo23-ku_2022_3dtiles%20_1_1_op_bldg_13101_chiyoda-ku_lod1/tileset.json";
