+++ web/e2e/project/utils.ts
@@ -2,25 +2,6 @@ import { Page } from "@playwright/test";
 
 import { expect } from "@reearth-cms/e2e/utils";
 
-export async function createProject(page: Page) {
-  await page.getByRole("button", { name: "plus New Project" }).first().click();
-  await page.getByRole("dialog").locator("#name").click();
-  await page.getByRole("dialog").locator("#name").fill("e2e project name");
-  await page.getByLabel("Project alias").click();
-  await page.getByLabel("Project alias").fill("e2e-project-alias");
-  await page.getByLabel("Project description").click();
-  await page.getByLabel("Project description").fill("e2e project description");
-  await page.getByRole("button", { name: "OK" }).click();
-  await page.getByText("e2e project name", { exact: true }).click();
-}
-
-export async function deleteProject(page: Page) {
-  await page.getByText("Settings").first().click();
-  await page.getByRole("button", { name: "Delete Project" }).click();
-  await page.getByRole("button", { name: "OK" }).click();
-  await expect(page.getByRole("alert").last()).toContainText("Successfully deleted project!");
-}
-
 export async function createModel(page: Page) {
   await page.getByText("Schema").click();
   await page.getByRole("button", { name: "plus Add" }).first().click();
